Fixed binary writing. Error management function add. TODO: ERROR TEST NORME LEAKS

+++ includes/op.c
@@ -12,7 +12,7 @@
 
 #include "asm.h"
 
-size_t g_size = 0;
+unsigned g_size = 0;
 
 int    g_count = 0;
